media upload stuff

+++ app/models/Media.java
@@ -1,6 +1,7 @@
 package models;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.text.SimpleDateFormat;
 import java.util.*;
 
@@ -19,7 +20,10 @@ import controllers.routes;
 
 import play.db.jpa.*;
 import play.mvc.Result;
+import scala.reflect.io.FileOperationException;
 import views.html.Course.view;
+
+import java.util.UUID;
 
 @Entity
 @SequenceGenerator(name = "default_seq", sequenceName = "media_seq")
@@ -48,6 +52,9 @@ public class Media extends BaseModel {
 	@ManyToOne
 	public Course course;
 	
+	@ManyToOne
+	public Account owner;
+	
 	@Transient
 	public File file;
 	
@@ -57,15 +64,29 @@ public class Media extends BaseModel {
 	    String path = Play.application().path().toString();
 	    String relPath = Play.application().configuration().getString("media.relativePath");
 		media.file = new File(path + "/" + relPath + "/" + media.url);
-		return media;
+		if(media.file.exists()) {
+			return media;
+		} else {
+			return null;
+		}
 	}
 	
 	@Override
 	public void create() {
+		
+	}
+	
+	public void create(String user) {
+		Account account = Account.findByEmail(user);
+		this.owner = account;
 		this.size = file.length();
-		this.url = this.createRelativeURL() + "/" + this.fileName;
-		this.createFile();
-		JPA.em().persist(this);
+		this.url = this.createRelativeURL() + "/" + this.getUniqueFileName(this.fileName);
+		try {
+			this.createFile();
+			JPA.em().persist(this);
+		} catch (FileOperationException e) {
+			throw e;
+		}
 	}
 	
 	@Override
@@ -78,13 +99,23 @@ public class Media extends BaseModel {
 	public void delete() {
 		JPA.em().remove(this);
 	}
+	
+	private String getUniqueFileName(String fileName) {
+		return UUID.randomUUID().toString() + '_' + fileName;
+	}
 		
-	private void createFile() {
+	private void createFile() throws FileOperationException {
 	    String path = Play.application().path().toString();
 	    String relPath = Play.application().configuration().getString("media.relativePath");
 	    File newFile = new File(path + "/" + relPath + "/" + this.url);
+	    if(newFile.exists()){
+	    	throw new FileOperationException("File exists already");
+	    }
 	    newFile.getParentFile().mkdirs();
 	    this.file.renameTo(newFile);
+	    if(!newFile.exists()) {
+	    	throw new FileOperationException("Could not upload file");
+	    }
 	}
 	
 	private String createRelativeURL() {
